backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
httpx